requirements.txt
bleak==0.21.1
opuslib==3.0.1
requests==2.31.0
python-dotenv==1.0.1